requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==22.2.0
build==0.10.0
CacheControl==0.12.11
cachetools==5.3.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
crashtest==0.4.1
cryptography==39.0.2
discord==2.3.2
discord.py==2.3.2
distlib==0.3.6
dulwich==0.21.3
filelock==3.9.0
firebase-admin==6.4.0
frozenlist==1.4.1
google-api-core==2.17.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.5
mysql-connector-python==8.3.0
packaging==23.0
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
rapidfuzz==2.13.7
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.0.post1
tomli==2.0.1
tomlkit==0.11.6
trove-classifiers==2023.3.9
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
webencodings==0.5.1
yarl==1.9.4
zipp==3.15.0
